Almost finished working GUI from linux
Require the code from the multigrid and extra graphs, added to be complete. Graph width needs to be increased after compiled. Dead pixels creating ghost plots are a problem. Change axis to (0,0) in middle, with drawn on in pixmap starting grid.

+++ sor.cpp
@@ -11,7 +11,7 @@ void MainWindow::SOR()
         int m=0;
         int i=0;
         int j=0;
-        float w = 1.6;
+        int w = 1.6;
         int p =0.99999;
         int c;
